+++ src/main/java/com/example/demo/base/event/listen/BasketEventListener.java
@@ -1,6 +1,6 @@
 package com.example.demo.base.event.listen;
 
-import com.example.demo.boundedContext.product.dto.Basket;
+import com.example.demo.boundedContext.product.entity.Basket;
 import com.example.demo.boundedContext.product.event.DeleteBasket;
 import com.example.demo.boundedContext.product.repository.BasketRepository;
 import lombok.RequiredArgsConstructor;
