+++ src/main/java/com/example/demo/boundedContext/member/controller/MemberController.java
@@ -8,7 +8,7 @@ import com.example.demo.boundedContext.member.service.MemberService;
 import com.example.demo.boundedContext.member.service.TestService;
 import com.example.demo.boundedContext.order.entity.Order;
 import com.example.demo.boundedContext.order.service.OrderService;
-import com.example.demo.boundedContext.product.dto.Basket;
+import com.example.demo.boundedContext.product.entity.Basket;
 import com.example.demo.boundedContext.product.repository.BasketRepository;
 import com.example.demo.util.rq.Rq;
 import com.nimbusds.oauth2.sdk.util.StringUtils;
