+++ src/main/java/com/example/demo/boundedContext/product/repository/BasketRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.boundedContext.product.repository;
 
-import com.example.demo.boundedContext.product.dto.Basket;
+import com.example.demo.boundedContext.product.entity.Basket;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
